pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fish-farm-env"
version = "1.0.0"
description = "Fish Farm OpenEnv — AI agent manages a Nile Tilapia RAS aquaculture system"
requires-python = ">=3.10"
authors = [
    {name = "Rahul Rajpurohit", email = "rahulrajpurohit2024@gmail.com"},
]
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
baseline = ["openai>=1.0.0"]
training = ["trl>=0.17.0", "transformers>=4.40.0", "datasets>=2.18.0", "accelerate>=0.28.0"]
dev = ["pytest>=7.0.0", "black", "ruff"]

[project.scripts]
server = "uvicorn:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.setuptools.packages.find]
where = ["src"]
